+++ src/cpp/rtps/messages/submessages/GapMsg.hpp
@@ -76,7 +76,7 @@ bool RTPSMessageCreator::addSubmessageGap(
     CDRMessage::addSequenceNumberSet(msg, &seqNumList);
 
     //TODO(Ricardo) Improve.
-    submessage_size = msg->pos - position_size_count_size;
+    submessage_size = uint16_t(msg->pos - position_size_count_size);
     octet* o= (octet*)&submessage_size;
     if(msg->msg_endian == DEFAULT_ENDIAN)
     {
